fix(openiddict): resolve all /audit findings (2 ARCHITECTURE + 4 CONVENTION/CLEANUP)
ARCHITECTURE fixes:
- Add [DependsOn(typeof(GranitEncryptionModule))] to EF Core module
  (direct dependency for key rotation IStringEncryptionService)
- Add [DependsOn(typeof(GranitSettingsModule))] to BackgroundJobs module
  (direct dependency for idle session ISettingProvider)

CONVENTION fixes:
- BackToImpersonator: .Produces<ImpersonationResult>() instead of .Produces(200)
- pt-BR.json: add missing PermissionGroup:OpenIddict key (24/24 keys)

CLEANUP fixes:
- Move AdminUserCreateRequest to Dtos/AdminUserCreateRequest.cs
- Move PasskeyRegistrationRequest + PasskeyRenameRequest to Dtos/PasskeyRequests.cs
- Update validator usings to reference Dtos namespace

All 69 tests pass. Build clean.
