+++ src/Granit.OpenIddict.Endpoints/Endpoints/AccountPasskeyEndpoints.cs
@@ -1,3 +1,4 @@
+using Granit.OpenIddict.Endpoints.Dtos;
 using Granit.OpenIddict.Services;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
@@ -149,12 +150,3 @@ internal static class AccountPasskeyEndpoints
         }
     }
 }
-
-/// <summary>Request DTO for passkey registration completion.</summary>
-/// <param name="CredentialJson">The WebAuthn AuthenticatorAttestationResponse JSON.</param>
-/// <param name="Name">Optional friendly name for the passkey.</param>
-public sealed record PasskeyRegistrationRequest(string CredentialJson, string? Name);
-
-/// <summary>Request DTO for passkey rename.</summary>
-/// <param name="Name">The new friendly name (max 100 chars).</param>
-public sealed record PasskeyRenameRequest(string Name);
