+++ src/Granit.OpenIddict.Endpoints/Endpoints/AdminUserEndpoints.cs
@@ -1,5 +1,6 @@
 using Granit.Identity;
 using Granit.Identity.Models;
+using Granit.OpenIddict.Endpoints.Dtos;
 using Granit.OpenIddict.Permissions;
 using Granit.OpenIddict.Services;
 using Microsoft.AspNetCore.Builder;
@@ -149,16 +150,3 @@ internal static class AdminUserEndpoints
         return TypedResults.Ok(result);
     }
 }
-
-/// <summary>Request DTO for admin user creation.</summary>
-/// <param name="Email">The user's email address.</param>
-/// <param name="FirstName">Optional first name.</param>
-/// <param name="LastName">Optional last name.</param>
-/// <param name="TemporaryPassword">Optional temporary password (forces change on next login).</param>
-#pragma warning disable GRSEC003 // DTO property name, not a secret
-public sealed record AdminUserCreateRequest(
-    string Email,
-    string? FirstName,
-    string? LastName,
-    string? TemporaryPassword);
-#pragma warning restore GRSEC003
